reupload

+++ pages/index.php
@@ -20,11 +20,11 @@ $workers = $stmt->fetchAll(PDO::FETCH_ASSOC);
 </head>
 
 <body>
+    <section class="container-fluid bg-dark d-flex justify-content-between py-3 px-3 mb-5">
+        <h2 class="text-white">All Users</h2>
+        <a href="add.php" class="text-primary h1"><i class='bx bx-user-plus'></i></a>
+    </section>
     <div class="container ">
-        <section class="d-flex justify-content-between mt-5">
-            <h2>All Users</h2>
-            <a href="add.php" class="text-primary h1"><i class='bx bx-user-plus'></i></a>
-        </section>
         <hr>
 
         <div
@@ -33,7 +33,7 @@ $workers = $stmt->fetchAll(PDO::FETCH_ASSOC);
                 class="table table-striped table-hover table-borderless  align-middle">
                 <thead class="table-light">
                     <tr>
-                        <th>#</th>
+                        <th>Photo</th>
                         <th>Name</th>
                         <th>Email</th>
                         <th>Phone</th>
@@ -42,16 +42,16 @@ $workers = $stmt->fetchAll(PDO::FETCH_ASSOC);
                 </thead>
                 <tbody>
                     <?php foreach ($workers as $worker): ?>
-                    <tr>
-                        <td><?= $worker['worker_id'] ?></td>
-                        <td><?= $worker['name'] ?></td>
-                        <td><?= $worker['email'] ?></td>
-                        <td><?= $worker['phone'] ?></td>
-                        <td>
-                            <a href="delete.php?id=<?= $worker['worker_id'] ?>" class="h3 text-danger me-3"><i class='bx bx-trash'></i></a>
-                            <a href="update.php?id=<?= $worker['worker_id'] ?>" class="h3 text-success"><i class='bx bx-edit' ></i></a>
-                        </td>
-                    </tr>
+                        <tr>
+                            <td><img src="<?= $worker['photo'] ?>" alt="<?= $worker['photo'] ?>" class="img-thumbnail w-25"></td>
+                            <td><?= $worker['name'] ?></td>
+                            <td><?= $worker['email'] ?></td>
+                            <td><?= $worker['phone'] ?></td>
+                            <td>
+                                <a href="delete.php?id=<?= $worker['worker_id'] ?>" class="h3 text-danger me-3" onclick="return confirm('you really want to delete this user?')"><i class='bx bx-trash'></i></a>
+                                <a href="update.php?id=<?= $worker['worker_id'] ?>" class="h3 text-success"><i class='bx bx-edit'></i></a>
+                            </td>
+                        </tr>
                     <?php endforeach ?>
                 </tbody>
             </table>
